requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
django-redis>=5.4
gunicorn>=20.1.0
python-dotenv>=1.0.0
django-environ
